requirements.txt
streamlit
opencv-python
numpy
tensorflow
google-generativeai
python-dotenv
